fix: pass the char, not its int cast, to convertHexDigit in parseInt
parseInt already had the character as a char, but called
convertHexDigit's int overload via an unnecessary (int) cast,
whose error message renders the numeric code point instead of
the character - "'71' is not a valid hex digit" instead of "'G'
is not a valid hex digit". Dropping the cast resolves to the char
overload instead, with no change to the valid-digit behavior.

+++ src/main/java/spark/utils/urldecoding/TypeUtil.java
@@ -65,7 +65,7 @@ public class TypeUtil {
         for (int i = 0; i < length; i++) {
             char c = s.charAt(offset + i);
 
-            int digit = convertHexDigit((int) c);
+            int digit = convertHexDigit(c);
             if (digit < 0 || digit >= base) {
                 throw new NumberFormatException("'" + c + "' at index " + (offset + i)
                         + " is not a valid base-" + base + " digit (in \"" + s + "\")");
